Update dependencies and restructure file tree
